Expand tests and move amp_theme logic into its own class
